gemini_agent/requirements.txt
google-generativeai>=0.8.0
pyaudio>=0.2.14
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
